chore: reorganized modal pages
