+++ frontend/src/components/question/selectors/SelectCategoriesInput.tsx
